requirements.txt
asgiref==3.5.0
astroid==2.11.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
dill==0.3.4
Django==4.0.3
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.1
psycopg2-binary
pylint==2.13.5
pymongo==4.1.0
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
wrapt==1.14.0
